Fix: Fix the output writing issue with python3
Python 3 was nagging about the wrong type of argument: "write() argument
must be str, not bytes." This is because the FileType of argparse was
using the binary mode to create the file handle, which just isn't the
right thing when printing strings. Additionally, stdout was using the
wrong format as well, which is resolved by using detach().